+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-10 16:44:29
-// Commit: feat(perf): add bi-directional event bus handler for perf
+// Updated at: 2026-08-10 19:41:26
+// Commit: feat(perf): implement responsive perf view component
 
 class perfComponentState {
-  final String commitRef = '3503';
-  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
-  final DateTime timestamp = DateTime.parse('2026-08-10T16:44:29+05:00');
+  final String commitRef = '3505';
+  final String lastMessage = 'feat(perf): implement responsive perf view component';
+  final DateTime timestamp = DateTime.parse('2026-08-10T19:41:26+05:00');
 }